Já escreve, ainda não apaga

+++ proj/headers/graphic.h
@@ -3,7 +3,7 @@
 
 #include <lcom/lcf.h>
 #include <machine/int86.h>
-#include <vbe.h>
+#include <../headers/vbe.h>
 
 int graphic_get_mode_info(uint16_t mode, vbe_mode_info_t *info);
 
@@ -17,7 +17,7 @@ int graphic_xpm(xpm_map_t xpm, uint16_t x, uint16_t y, bool trans);
 
 int graphic_cntrl_info(vg_vbe_contr_info_t *info);
 
-int32_t get_char_xpm(char c);
+uint8_t get_char_xpm(char c);
 
 #endif /* _LCOM_VG_H */
